+++ test/test_auth/test_access_control.py
@@ -7,7 +7,7 @@ from asab.web.auth.authorization import (
 )
 
 
-class TestCommonUser(unittest.TestCase):
+class TestAccessControlCommonUser(unittest.TestCase):
 	Claims = {
 		"resources": {
 			"*": [
@@ -119,6 +119,120 @@ class TestCommonUser(unittest.TestCase):
 
 	def test_superuser(self):
 		self.assertFalse(
+			is_superuser(
+				claims=self.Claims,
+			)
+		)
+
+
+class TestAccessControlSuperuser(unittest.TestCase):
+	Claims = {
+		"resources": {
+			"*": [
+				"authz:superuser",
+			],
+		},
+	}
+
+	def test_tenant_resource_access(self):
+		self.assertTrue(
+			has_resource_access(
+				claims=self.Claims,
+				resources=["read", "write", "delete"],
+				tenant="good",
+			)
+		)
+		self.assertTrue(
+			has_resource_access(
+				claims=self.Claims,
+				resources=["read", "write"],
+				tenant="good",
+			)
+		)
+		self.assertTrue(
+			has_resource_access(
+				claims=self.Claims,
+				resources=["read"],
+				tenant="good",
+			)
+		)
+		self.assertTrue(
+			has_resource_access(
+				claims=self.Claims,
+				resources=["read", "write"],
+				tenant="bad",
+			)
+		)
+		self.assertTrue(
+			has_resource_access(
+				claims=self.Claims,
+				resources=["read"],
+				tenant="bad",
+			)
+		)
+		self.assertRaises(
+			ValueError,
+			lambda: has_resource_access(
+				claims=self.Claims,
+				resources=["read"],
+				tenant="*",
+			)
+		)
+
+	def test_global_resource_access(self):
+		self.assertTrue(
+			has_resource_access(
+				claims=self.Claims,
+				resources=["read", "write"],
+				tenant=None,
+			)
+		)
+		self.assertTrue(
+			has_resource_access(
+				claims=self.Claims,
+				resources=["read"],
+				tenant=None,
+			)
+		)
+		self.assertRaises(
+			ValueError,
+			lambda: has_resource_access(
+				claims=self.Claims,
+				resources=["read"],
+				tenant="*",
+			)
+		)
+
+	def test_tenant_access(self):
+		self.assertTrue(
+			has_tenant_access(
+				claims=self.Claims,
+				tenant="good",
+			)
+		)
+		self.assertTrue(
+			has_tenant_access(
+				claims=self.Claims,
+				tenant="bad",
+			)
+		)
+		self.assertRaises(
+			ValueError,
+			lambda: has_tenant_access(
+				claims=self.Claims,
+				tenant="*",
+			)
+		)
+		self.assertRaises(
+			ValueError,
+			lambda: has_tenant_access(
+				claims=self.Claims,
+				tenant=None,
+			)
+		)
+
+	def test_superuser(self):
+		self.assertTrue(
 			is_superuser(
 				claims=self.Claims,
 			)
